rewrites for 3.6
