+++ EventHubExample.sln
@@ -1,11 +1,11 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2013
-VisualStudioVersion = 12.0.31101.0
+# Visual Studio 15
+VisualStudioVersion = 15.0.26430.16
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EventHubWriter", "EventHubWriter\EventHubWriter.csproj", "{63096311-06E9-4EC1-8D0F-C930BF922A70}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EventHubWriter", "EventHubWriter\EventHubWriter.csproj", "{B8891078-5E34-4FBE-A7A0-A79B80232E37}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EventHubReader", "EventHubReader\EventHubReader.csproj", "{45EA7BEF-B033-4DDC-93EE-F659B3669E62}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EventHubReader", "EventHubReader\EventHubReader.csproj", "{0752E859-14FF-4B40-829C-9A92356C5E22}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -13,14 +13,14 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{63096311-06E9-4EC1-8D0F-C930BF922A70}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{63096311-06E9-4EC1-8D0F-C930BF922A70}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{63096311-06E9-4EC1-8D0F-C930BF922A70}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{63096311-06E9-4EC1-8D0F-C930BF922A70}.Release|Any CPU.Build.0 = Release|Any CPU
-		{45EA7BEF-B033-4DDC-93EE-F659B3669E62}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{45EA7BEF-B033-4DDC-93EE-F659B3669E62}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{45EA7BEF-B033-4DDC-93EE-F659B3669E62}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{45EA7BEF-B033-4DDC-93EE-F659B3669E62}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B8891078-5E34-4FBE-A7A0-A79B80232E37}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B8891078-5E34-4FBE-A7A0-A79B80232E37}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B8891078-5E34-4FBE-A7A0-A79B80232E37}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B8891078-5E34-4FBE-A7A0-A79B80232E37}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0752E859-14FF-4B40-829C-9A92356C5E22}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0752E859-14FF-4B40-829C-9A92356C5E22}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0752E859-14FF-4B40-829C-9A92356C5E22}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0752E859-14FF-4B40-829C-9A92356C5E22}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
