+++ EventHubReader/App.config
@@ -1,21 +1,13 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8" ?>
 <configuration>
-  <startup>
-    <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.5" />
-  </startup>
+    <startup>
+        <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.5" />
+    </startup>
   <appSettings>
-      <add key="EventHubNamespace" value="##TOBEFILLED##" />
-      <add key="EventHubEntityPath" value="##TOBEFILLED##" />
-      <add key="EventHubSharedAccessKeyName" value="reader" />
-      <add key="EventHubSharedAccessKey" value="##TOBEFILLED##" />
-      <add key="EventHubPartitions" value="2" />
+    <add key="EventHubNamespace" value="##TOBEFILLED##" />
+    <add key="EventHubEntityPath" value="##TOBEFILLED##" />
+    <add key="EventHubSharedAccessKeyName" value="##TOBEFILLED##" />
+    <add key="EventHubSharedAccessKey" value="##TOBEFILLED##" />
+    <add key="EventHubPartitions" value="2" />
   </appSettings>
-  <runtime>
-    <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
-      <dependentAssembly>
-        <assemblyIdentity name="Newtonsoft.Json" publicKeyToken="30ad4fe6b2a6aeed" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-9.0.0.0" newVersion="9.0.0.0" />
-      </dependentAssembly>
-    </assemblyBinding>
-  </runtime>
 </configuration>
