+++ EventHubReader/EventHubReader.csproj
@@ -1,16 +1,16 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="4.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <Import Project="..\packages\Microsoft.SCP.Net.SDK.1.0.0.1\build\Microsoft.SCP.Net.SDK.props" Condition="Exists('..\packages\Microsoft.SCP.Net.SDK.1.0.0.1\build\Microsoft.SCP.Net.SDK.props')" />
+  <Import Project="..\packages\Microsoft.SCP.Net.SDK.1.0.0.3\build\Microsoft.SCP.Net.SDK.props" Condition="Exists('..\packages\Microsoft.SCP.Net.SDK.1.0.0.3\build\Microsoft.SCP.Net.SDK.props')" />
   <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
     <ProjectTypeGuids>{12D26815-A329-4BD9-922E-4A35B783B8AA};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
-    <ProjectGuid>{45EA7BEF-B033-4DDC-93EE-F659B3669E62}</ProjectGuid>
+    <ProjectGuid>{0752E859-14FF-4B40-829C-9A92356C5E22}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>AzureEventHubsReaderStormApplication</RootNamespace>
-    <AssemblyName>AzureEventHubsReaderStormApplication</AssemblyName>
+    <RootNamespace>EventHubReader</RootNamespace>
+    <AssemblyName>EventHubReader</AssemblyName>
     <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <NuGetPackageImportStamp>
@@ -36,21 +36,18 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
-    <Reference Include="log4net, Version=1.2.10.0, Culture=neutral, PublicKeyToken=1b44e1d426115821, processorArchitecture=MSIL">
-      <HintPath>..\packages\Microsoft.SCP.Net.SDK.1.0.0.1\lib\net45\log4net.dll</HintPath>
-      <Private>True</Private>
+    <Reference Include="log4net, Version=1.2.15.0, Culture=neutral, PublicKeyToken=669e0ddf0bb1aa2a, processorArchitecture=MSIL">
+      <HintPath>..\packages\Microsoft.SCP.Net.SDK.1.0.0.3\lib\net45\log4net.dll</HintPath>
     </Reference>
     <Reference Include="Microsoft.SCPNet, Version=1.0.0.0, Culture=neutral, processorArchitecture=MSIL">
-      <HintPath>..\packages\Microsoft.SCP.Net.SDK.1.0.0.1\lib\net45\Microsoft.SCPNet.dll</HintPath>
-      <Private>True</Private>
+      <HintPath>..\packages\Microsoft.SCP.Net.SDK.1.0.0.3\lib\net45\Microsoft.SCPNet.dll</HintPath>
     </Reference>
-    <Reference Include="Newtonsoft.Json, Version=9.0.0.0, Culture=neutral, PublicKeyToken=30ad4fe6b2a6aeed, processorArchitecture=MSIL">
-      <HintPath>..\packages\Newtonsoft.Json.9.0.1\lib\net45\Newtonsoft.Json.dll</HintPath>
+    <Reference Include="Newtonsoft.Json, Version=6.0.0.0, Culture=neutral, PublicKeyToken=30ad4fe6b2a6aeed, processorArchitecture=MSIL">
+      <HintPath>..\packages\Newtonsoft.Json.6.0.8\lib\net45\Newtonsoft.Json.dll</HintPath>
       <Private>True</Private>
     </Reference>
     <Reference Include="SCPHost, Version=1.0.0.0, Culture=neutral, processorArchitecture=MSIL">
-      <HintPath>..\packages\Microsoft.SCP.Net.SDK.1.0.0.1\lib\net45\SCPHost.exe</HintPath>
-      <Private>True</Private>
+      <HintPath>..\packages\Microsoft.SCP.Net.SDK.1.0.0.3\lib\net45\SCPHost.exe</HintPath>
     </Reference>
     <Reference Include="System" />
     <Reference Include="System.Configuration" />
@@ -61,38 +58,30 @@
     <Reference Include="System.Data" />
     <Reference Include="System.Xml" />
     <Reference Include="Thrift, Version=1.0.0.1, Culture=neutral, processorArchitecture=MSIL">
-      <HintPath>..\packages\Microsoft.SCP.Net.SDK.1.0.0.1\lib\net45\Thrift.dll</HintPath>
-      <Private>True</Private>
+      <HintPath>..\packages\Microsoft.SCP.Net.SDK.1.0.0.3\lib\net45\Thrift.dll</HintPath>
     </Reference>
     <Reference Include="ZooKeeperNet, Version=3.4.6.1, Culture=neutral, PublicKeyToken=fefd2c046da35b56, processorArchitecture=MSIL">
-      <HintPath>..\packages\Microsoft.SCP.Net.SDK.1.0.0.1\lib\net45\ZooKeeperNet.dll</HintPath>
-      <Private>True</Private>
+      <HintPath>..\packages\Microsoft.SCP.Net.SDK.1.0.0.3\lib\net45\ZooKeeperNet.dll</HintPath>
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="EventHubsReaderTopology.cs" />
     <Compile Include="LogBolt.cs" />
+    <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
-    <None Include="App.config">
-      <SubType>Designer</SubType>
-    </None>
+    <None Include="App.config" />
     <None Include="packages.config" />
-    <None Include="SubmitConfig.xml" />
-    <None Include="Microsoft.SCPNet.dll.config">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
     <PropertyGroup>
       <ErrorText>This project references NuGet package(s) that are missing on this computer. Use NuGet Package Restore to download them.  For more information, see http://go.microsoft.com/fwlink/?LinkID=322105. The missing file is {0}.</ErrorText>
     </PropertyGroup>
-    <Error Condition="!Exists('..\packages\Microsoft.SCP.Net.SDK.1.0.0.1\build\Microsoft.SCP.Net.SDK.props')" Text="$([System.String]::Format('$(ErrorText)', '..\packages\Microsoft.SCP.Net.SDK.1.0.0.1\build\Microsoft.SCP.Net.SDK.props'))" />
-    <Error Condition="!Exists('..\packages\Microsoft.SCP.Net.SDK.1.0.0.1\build\Microsoft.SCP.Net.SDK.targets')" Text="$([System.String]::Format('$(ErrorText)', '..\packages\Microsoft.SCP.Net.SDK.1.0.0.1\build\Microsoft.SCP.Net.SDK.targets'))" />
+    <Error Condition="!Exists('..\packages\Microsoft.SCP.Net.SDK.1.0.0.3\build\Microsoft.SCP.Net.SDK.props')" Text="$([System.String]::Format('$(ErrorText)', '..\packages\Microsoft.SCP.Net.SDK.1.0.0.3\build\Microsoft.SCP.Net.SDK.props'))" />
+    <Error Condition="!Exists('..\packages\Microsoft.SCP.Net.SDK.1.0.0.3\build\Microsoft.SCP.Net.SDK.targets')" Text="$([System.String]::Format('$(ErrorText)', '..\packages\Microsoft.SCP.Net.SDK.1.0.0.3\build\Microsoft.SCP.Net.SDK.targets'))" />
   </Target>
-  <Import Project="..\packages\Microsoft.SCP.Net.SDK.1.0.0.1\build\Microsoft.SCP.Net.SDK.targets" Condition="Exists('..\packages\Microsoft.SCP.Net.SDK.1.0.0.1\build\Microsoft.SCP.Net.SDK.targets')" />
+  <Import Project="..\packages\Microsoft.SCP.Net.SDK.1.0.0.3\build\Microsoft.SCP.Net.SDK.targets" Condition="Exists('..\packages\Microsoft.SCP.Net.SDK.1.0.0.3\build\Microsoft.SCP.Net.SDK.targets')" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
   <Target Name="BeforeBuild">
