--- EventHubReader/Microsoft.SCPNet.dll.config
@@ -1,91 +1,0 @@
-﻿<?xml version="1.0"?>
-<configuration>
-  <configSections>
-    <section name="log4net" type="log4net.Config.Log4NetConfigurationSectionHandler, log4net"/>
-    <section name="SCPLogger" type="Microsoft.SCP.SCPLogger.Config.SCPLogger, Microsoft.SCPLogger"/>
-    <section name="SCPPerfCounter" type="Microsoft.SCP.SCPPerfCounter.Config.SCPPerfCounterConfig, SCPPerfCounter"/>
-  </configSections>
-  <SCPLogger>
-    <LoggerFactories>
-      <LoggerFactory type="SCPLogger.Log4net.Log4netLoggerFactory, SCPLogger.Log4net" />
-    </LoggerFactories>
-  </SCPLogger>
-  <SCPPerfCounter type="Microsoft.SCP.SCPPerfCounter.NullPerfCounterFactory, SCPPerfCounter">
-  </SCPPerfCounter>
-  <appSettings>
-    <add key="BatchSize" value="3000" />
-  </appSettings>
-  <log4net>
-    <!--Console appenders for writing to stdout & stderr that gets routed to Storm Worker JVM logs-->
-    <appender name="ConsoleAppenderError" type="log4net.Appender.ConsoleAppender">
-      <threshold value="ERROR"/>
-      <target value="Console.Error" />
-      <layout type="log4net.Layout.PatternLayout">
-        <conversionPattern value="%date [%thread] %-5level %logger [%ndc] - %message%newline" />
-      </layout>
-    </appender>
-    <appender name="ConsoleAppender" type="log4net.Appender.ConsoleAppender">
-      <threshold value="INFO"/>
-      <target value="Console.Out" />
-      <layout type="log4net.Layout.PatternLayout">
-        <conversionPattern value="%date [%thread] %-5level %logger [%ndc] - %message%newline" />
-      </layout>
-    </appender>
-    <appender name="ErrorAppender" type="log4net.Appender.RollingFileAppender">
-      <file type="log4net.Util.PatternString">
-        <conversionPattern value="%env{STORM_HOME}\logs\scp\%property{LogName}_%property{ProcessId}_error_" />
-      </file>
-      <param name="AppendToFile" value="true" />
-      <param name="MaxSizeRollBackups" value="10" />
-      <param name="MaxFileSize" value="10240000" />
-      <param name="StaticLogFileName" value="false" />
-      <param name="DatePattern" value="yyyy_MM_dd&quot;.log&quot;" />
-      <param name="RollingStyle" value="Composite" />
-      <param name="PreserveLogFileNameExtension" value="true" />
-      <layout type="log4net.Layout.PatternLayout">
-        <conversionPattern value="%date [%thread] %-5level %logger [%ndc] - %message%newline" />
-      </layout>
-    </appender>
-    <appender name="RollingFileAppender" type="log4net.Appender.RollingFileAppender">
-      <file type="log4net.Util.PatternString">
-        <conversionPattern value="%env{STORM_HOME}\logs\scp\%property{LogName}_%property{ProcessId}_" />
-      </file>
-      <param name="AppendToFile" value="true" />
-      <param name="MaxSizeRollBackups" value="10" />
-      <param name="MaxFileSize" value="102400000" />
-      <param name="StaticLogFileName" value="false" />
-      <param name="DatePattern" value="yyyy_MM_dd&quot;.log&quot;" />
-      <param name="RollingStyle" value="Composite" />
-      <param name="PreserveLogFileNameExtension" value="true" />
-      <layout type="log4net.Layout.PatternLayout">
-        <conversionPattern value="%date [%thread] %-5level %logger [%ndc] - %message%newline" />
-      </layout>
-    </appender>
-    <appender name="HdInsightEtwAppender" type="SCPLogger.Log4net.EtwAppender.HdInsightEtwAppender, SCPLogger.Log4net.EtwAppender">
-      <ComponentName value="storm-scpnet" />
-      <layout type="log4net.Layout.PatternLayout">
-        <conversionPattern value="%date [%thread] %-5level %F:%L - %message%newline" />
-      </layout>
-    </appender>
-    <appender name="HdInsightFilterEtwAppender" type="SCPLogger.Log4net.EtwAppender.HdInsightFilterEtwAppender, SCPLogger.Log4net.EtwAppender">
-      <threshold value="WARN"/>
-      <ComponentName value="storm-scpnet" />
-      <layout type="log4net.Layout.PatternLayout">
-        <conversionPattern value="%date [%thread] %-5level %F:%L - %message%newline" />
-      </layout>
-    </appender>
-
-    <root>
-      <level value="INFO" />
-      <appender-ref ref="ConsoleAppenderError" />
-      <appender-ref ref="ConsoleAppender" />
-      <appender-ref ref="RollingFileAppender" />
-      <appender-ref ref="HdInsightEtwAppender" />
-      <appender-ref ref="HdInsightFilterEtwAppender" />
-    </root>
-  </log4net>
-  <startup useLegacyV2RuntimeActivationPolicy="true">
-    <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.0" />
-    <requiredRuntime version="v4.0.20506" />
-  </startup>
-</configuration>
