+++ EventHubWriter/App.config
@@ -1,13 +1,13 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <configuration>
-    <startup> 
-        <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.5" />
-    </startup>
+  <startup>
+    <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.5" />
+  </startup>
   <appSettings>
-    <add key="EventHubPolicyName" value="writer"/>
+    <add key="EventHubPolicyName" value="##TOBEFILLED##" />
     <add key="EventHubPolicyKey" value="##TOBEFILLED##" />
     <add key="EventHubNamespace" value="##TOBEFILLED##" />
     <add key="EventHubName" value="##TOBEFILLED##" />
-    <add key="EventHubPartitionCount" value="2"/>
+    <add key="EventHubPartitionCount" value="2" />
   </appSettings>
 </configuration>
