--- EventHubWriter/Microsoft.SCP.Net.SDK/sdk/SCPLogger.Log4net.EtwAppender.dll.config
@@ -1,29 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<configuration>
-  <!--<configSections>
-    <section name="log4net" type="log4net.Config.Log4NetConfigurationSectionHandler, log4net" />
-  </configSections>
-  <log4net>
-    <appender name="HdInsightEtwAppender" type="SCPLogger.Log4net.EtwAppender.HdInsightEtwAppender, SCPLogger.Log4net.EtwAppender">
-      <ComponentName value="storm-scpnet" />
-      <layout type="log4net.Layout.PatternLayout">
-        <conversionPattern value="%date [%thread] %-5level %F:%L - %message%newline" />
-      </layout>
-    </appender>
-    <appender name="HdInsightFilterEtwAppender" type="SCPLogger.Log4net.EtwAppender.HdInsightFilterEtwAppender, SCPLogger.Log4net.EtwAppender">
-      <threshold value="WARN"/>
-      <ComponentName value="storm-scpnet" />
-      <layout type="log4net.Layout.PatternLayout">
-        <conversionPattern value="%date [%thread] %-5level %F:%L - %message%newline" />
-      </layout>
-    </appender>
-
-    <logger name="SCPLogger.Log4net.EtwAppender.Test">
-      <level value="INFO"/>
-      <appender-ref ref="HdInsightEtwAppender" />
-      <appender-ref ref="HdInsightFilterEtwAppender" />
-    </logger>
-
-    <root></root>
-  </log4net>-->
-</configuration>
