--- EventHubWriter/Microsoft.SCP.Net.SDK/sdk/scp.log.config
@@ -1,32 +1,0 @@
-<?xml version="1.0"?>
-<configuration>
-  <configSections>
-    <section name="log4net" type="log4net.Config.Log4NetConfigurationSectionHandler, log4net"/>
-  </configSections>
-  <log4net>
-    <appender name="ConsoleAppender" type="log4net.Appender.ConsoleAppender">
-      <layout type="log4net.Layout.PatternLayout">
-        <conversionPattern value="%date [%thread] %-5level %logger - %message%newline"/>
-      </layout>
-    </appender>
-    <appender name="RollingFileAppender" type="log4net.Appender.RollingFileAppender">
-      <file type="log4net.Util.PatternString">
-        <conversionPattern value="scp.worker-%processid"/>
-      </file>
-      <param name="AppendToFile" value="true"/>
-      <param name="MaxSizeRollBackups" value="100"/>
-      <param name="MaxFileSize" value="10240000"/>
-      <param name="StaticLogFileName" value="false"/>
-      <param name="DatePattern" value="yyyy_MM_dd&quot;.log&quot;"/>
-      <param name="RollingStyle" value="Composite"/>
-      <!--param name="PreserveLogFileNameExtension" value="true"/-->
-      <layout type="log4net.Layout.PatternLayout">
-        <conversionPattern value="%date [%thread] %-5level %logger - %message%newline"/>
-      </layout>
-    </appender>
-    <root>
-      <level value="ERROR"/>
-      <appender-ref ref="RollingFileAppender"/>
-    </root>
-  </log4net>
-<startup><supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.0"/></startup></configuration>
