--- EventHubWriter/Microsoft.SCP.Net.SDK/tools/Install.ps1
@@ -1,3 +1,0 @@
-param($installPath, $toolsPath, $package, $project)
-
-
