--- EventHubWriter/Microsoft.SCP.Net.SDK/tools/ScpWebApiClient.exe.config
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-  <appSettings>
-    <add key="SCPAPIURL" value=""/>
-    <add key="Username" value=""/>
-    <add key="Password" value=""/>
-    <add key="OutputResponse" value="true"/>
-  </appSettings>
-  <startup>
-    <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.5"/>
-  </startup>
-</configuration>
