--- EventHubWriter/Properties/Settings.settings
@@ -1,5 +1,0 @@
-﻿<?xml version='1.0' encoding='utf-8'?>
-<SettingsFile xmlns="http://schemas.microsoft.com/VisualStudio/2004/01/settings" CurrentProfile="(Default)">
-  <Profiles />
-  <Settings />
-</SettingsFile>
