+++ EventHubWriter/packages.config
@@ -1,5 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Microsoft.SCP.Net.SDK" version="1.0.0.1" targetFramework="net45" />
-  <package id="Newtonsoft.Json" version="9.0.1" targetFramework="net45" />
+  <package id="log4net" version="1.2.10" targetFramework="net45" />
+  <package id="Microsoft.SCP.Net.SDK" version="1.0.0.3" targetFramework="net45" />
+  <package id="Newtonsoft.Json" version="6.0.8" targetFramework="net45" />
+  <package id="ZooKeeper.Net" version="3.4.6.2" targetFramework="net45" />
 </packages>
